Refactor tagging test as per changes during merge master
